update docs + improve examples + delete unused code

+++ forcex/src/com/forcex/app/InputListener.java
@@ -3,5 +3,5 @@ package com.forcex.app;
 public interface InputListener {
 	public void onKeyEvent(byte key,boolean down);
 
-	public void onTouch(float x,float y,byte type,byte pointer);
+	public void onTouch(float x,float y,byte type, int pointer);
 }
